Move common code in commons module
