+++ Apromore-Calendar/src/main/resources/META-INF/spring/calendar-osgi-context.xml
@@ -30,6 +30,6 @@
    
   
     <osgi:reference id="customCalendarRepository" interface="org.apromore.dao.CustomCalendarRepository"/>
-  
+   <osgi:service ref="calendarService" interface="org.apromore.calendar.service.CustomCalendarService" auto-export="interfaces"/>
 
 </beans:beans>
