+++ Apromore-Calendar/src/test/java/org/apromore/calendar/service/DurationCalculationUnitTestWithHoliday.java
@@ -61,6 +61,8 @@ public class DurationCalculationUnitTestWithHoliday {
     assertThat(durationModel.getDuration()).isEqualTo(expected);
   }
   
+//    add test for midnight shift workers.
+  
 
 
 }
