+++ Apromore-Calendar/src/test/resources/database/test-config.properties
@@ -12,7 +12,8 @@ jpa.databasePlatform = org.eclipse.persistence.platform.database.H2Platform
 jpa.showSql          = false
 jpa.generateDDL      = false
 
-
+liquibase.jdbc.username = apromore
+liquibase.jdbc.password = MAcri
 
 # Database and JPA (MySQL)
